Map and Points

+++ app/src/main/java/com/zapir/ariadne/model/entity/Point.kt
@@ -2,6 +2,6 @@ package com.zapir.ariadne.model.entity
 
 data class Point(
         val name: String? = null,
-        val x: Double = 0.0,
-        val y: Double = 0.0
+        val x: Float = 0.0f,
+        val y: Float = 0.0f
 )
